[web_admin] リファクタリング

+++ business_and_realworld/website_admin/pages/api/articles/delete.ts
@@ -1,22 +1,26 @@
 import { auth0 } from "../../../lib/server/auth0";
 
 export default auth0.requireAuthentication(async (req, res) => {
-  const id = req.query.id;
+  const {
+    query: { id },
+    method,
+  } = req;
 
   if (typeof id !== "string") {
     res.status(404).end();
     return;
   }
 
-  try {
-    await fetch(`https://website-admin.microcms.io/api/v1/articles/${id}`, {
-      method: "DELETE",
-      headers: { "X-WRITE-API-KEY": process.env.X_WRITE_API_KEY || "" },
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(error.status || 500).end(error.message);
+  if (method !== "DELETE") {
+    res.setHeader("Allow", ["DELETE"]);
+    res.status(405).end();
+    return;
   }
+
+  await fetch(`https://website-admin.microcms.io/api/v1/articles/${id}`, {
+    method: "DELETE",
+    headers: { "X-WRITE-API-KEY": process.env.X_WRITE_API_KEY || "" },
+  });
 
   res.status(200).end();
 });
